requirements.txt
flask==3.0.0
pillow==10.4.0
transformers==4.44.2
torch==2.4.1
safetensors==0.4.4